Basic example code using pygame
Need to add a few more bits to the toolbox in order to understand
how the whole system works, but we have the basics of things in there
now.  Images, mouse motion, and cursor keys are probably the next big
things to get going; then it would be worth looking into rectangles and
the things that those provide.