Added some font/text rendering
Man, this stuff is amazingly simple.  Now I have text!  Next up is to
do some stuff with rectangles -- see if we can do something that makes a
box (geometry drawing primitives!) that will "hold up" the mouse a bit
so that it can't quite go where it wants.  Should be interesting.