World kinda works
Still have a bunch of reference counting problems here -- need to
make sure that we drop the running world when we get a new one.
Also need to do some other things to make stuff be consistent.  But
we definitely have a start!

Need to add modifier bits to the world table so that walls can be
impassable, doors can be links, and so forth...