Fleshed out Hero, added Vec
If you're in worldmap mode, you can now steer around a purple box.
The controls here, while smooth, are extremely frustrating -- I need
to figure out how to model the behavior that I want the right way,
and I clearly don't have that as of yet.  But, things move, and the
math seems to be more or less working, so that's all good.