A bunch of weapon work
Added Sword class as starting point for a real weapon.  Lots of dangling
bits left to handle there at present, but the overall shape is starting
to become clear.  This might even work!  Still need to pound through
a whole pile of TODO entries in this file, but a bunch of the skeleton
is actually in place.