Minor changes
Getting my head back into the game here (har har) and added a few
comments about things for the Weapon class (should probably handle
input for all subclasses to make things easy) and added some stuff
to the plan so that I can track what's going on.

Enough of the Sword is in place now that it seems like I ought
to see about building an opponent and a means by which to place
them on the map and to make it so that combat begins when you
get close enough to them.  Whee!