Roughed in NPC class
Basic layout and interactions for the NPC class are now in place.
For now, I'm putting all of the interactivity decisions up into the
Hero class, which is likely where the world bound collision stuff
for the Hero will also go.  Seems like the most straightforward way
to handle things for now; may eventually need to do something that
involves more generic behaviors, but this should get us running.

Need to make a usable subclass of NPC that actually fights you and
whatnot.  Placeholder commented out code exists in Menu.py for the
place where we'd add one or two of those.

Inching closer to trying combat.