Goon added
This is a basic opponent.  The menu adds one to the world so that we
can experiment with combat.  Right now, we appear to enter combat
correctly, but then appear to be stuck since the hero doesn't have
a weapon, and thus doesn't have any kind of combat behavior.

Seems like the easy way to manage that is to have the Sword in fact
not be registered for update/render/handling, but instead, to have
the hero forward such calls along to the current active weapon.  Maybe
I'm wrong, but that seems like the logical thing to do.

Anyway, not really tested yet (no way to see progress, changes,
damage, etc.), but approximately in and working.