Combat functional!
Fixed up a bunch of random stuff this time through and got up
to the point that combat actually works.  Horray!  Still have
plenty of edges to clean up, but the basic system is now in
place.

Surface tagging for the world is the next big thing; that'll make
it so that we can (a) have an arena, and (b) do some more realistic
spawning of opponents.