World collision with Hero
The world object now tracks a list of rectangles that represent
walls (impassible areas).  The hero also now pays attention to
collision information with the world and doesn't allow motion into
the wall.  Altogether, this gives us basic world collision.

There seems to be a slight problem with my velocity change
computations, but this is relatively minor and should be easy
to correct for later on.  If even necessary.