Joystick support for menu
Fleshed out enough bits on the joystick class and its wiring into the
main game loop that it is now usable for navigation.  Next logical
step will be to add left stick player navigation, I think.

Also need to consider doing a branch of some kind at this point so
that I can produce the kind of game that I want for the girls, but
not completely toss the arithmetic notion on the floor.