Running example for hero with joystick control
Fixed up a bunch of the different cases in the code to deal at least
somewhat with having multiple heroes at the same time.  Tested with
one hero controlled via the joystick; may wish to open up the center
slop zones on the sticks, especially for the girls.  May also
consider applying a quadratic function to the system for determining
user input, so that you go relatively slowly outside of the zone
and go faster as you get further deflected.