Working gates
Worlds can now define gates properly, which allow warping from
one world to the next.  Added a second world file as an example
of the technology.

Also added a text file tracking things that need to be done
on carcar (carcar.txt) so that I have a simple todo list and
other scratch pad area.