Fixed up to basically work
The test cases we have (such as they are!) seem to be operating
correctly at this point. There is much that I do not remember at this
moment about how this is all supposed to fit together, so it's entirely
possible that there is code in here that's still broken and I just don't
know because it's not being exercised. In particular, I think I was
experimenting with something and don't actually make use of Goon objects
at the moment, so...yeah.