Finished prep for 2022-10-06
I think I have enough pieces in place that we may be able to get things
going for the project. We may need to poke around a bit to figure out a
reasonable font. We may also need to install YAML ourselves, but it
feels like maybe that's worth doing just to make it all go - can use
some of our powers from previous projects to maybe get that to operate,
woo!