Fix NPCs living across world boundaries, boot help
NPCs will now by default kill themselves when we switch worlds. It's a
function, so subclasses can override if needed. Additionally, sorted out
some stuff to try to get more immediate output when using boot.py so
that intermediate output can be observed while the game is running in
that mode, which should assist in debugging/testing scenarios.