HP bar display, in progress
Added basic HP bar to the hero. Basically works, although could use some
sprucing up to make the colors more interesting as damage is taken. Also
fixed several bugs where things didn't work, and improved the behavior
of the Episode menu (which was confusing -- it left itself pushed after
selecting a world, which was kind of goofy for subsequent uses of the
menu).