Sanitized update/render/handler priorities
Converted from silly number-based scheme to a set of Enums in Game.py
which just list out the various categories in order. Makes maintenance
much easier, and cleans up the Add/Remove interfaces a bit, at the cost
of requiring two functions to operate in a given system (e.g., OnX and
Xpri functions are now needed). This model should be much easier to
maintain moving forward, at the marginal cost of requiring the addition
of an enum type to each of the categories that an new object
participates in.