Added doors
These are functionally walls when they're closed, and are functionally
floor (no collision) when they're open. Can be authored in either state,
and has update logic to check for hero items that can cause them to
switch from the closed to the open state. The item checking logic is
probably incorrect vs. the functions that are currently exposed on Item
class in another branch, so we'll need to fix those up once we're all
there. Otherwise, functionality has been basically tested to confirm it
is working.