Added support for auto boss key
When a boss class object is killed, it now automatically adds an item to
the hero's inventory with the boss-key tag. We should be set therefore
to make various other things in the world respond appropriately such as
doors after the boss has been defeated.

Also fixed a bug where the boss was getting to update and render twice
each frame (!!) so we may need to adjust some timing numbers again.