Added attack damage to hero finder
This makes things a bit more interesting/challenging in cases where we
use these enemies. Numbers are tuned to be pretty low/slow, with the
notion that these are dangerous mostly in numbers, not individually,
although even two can be a bit of a challenge if you're limited in space
to manouver.